# docs/swagger.yaml
basePath: /api/v1
definitions:
  Content:
    properties:
      aspect_ratio:
        type: string
      category:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      play_time:
        type: string
      resolution:
        type: string
      tag:
        type: string
      thumbnail:
        type: string
      updatedAt:
        type: string
    required:
    - name
    - thumbnail
    type: object
  ErrorResponse:
    properties:
      code:
        type: string
      debug_message:
        type: string
    type: object
...
  Response:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /contents:
    get:
      consumes:
      - application/json
      description: content
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Content'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: content
      tags:
      - Content
    post:
      consumes:
      - application/json
      description: content
      parameters:
      - description: Body
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/Content'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: content
      tags:
      - Content
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"